(maint) Remove old CODEOWNERS file
This commit removes the old CODEOWNERS file. I didn't realize this existed when
adding the other one. All of Puppet's other repositories have the CODEOWNERS
file at the top level of the repo for easy findability, so that should replace
this one. This one also contained individual Puppet employees, rather than a
team, which is what we prefer.