(maint) Add CODEOWNERS (#77)
* (maint) Add CODEOWNERS

This commit replaces the OWNERS file with a CODEOWNERS file. This is a special
file that GitHub uses to automatically assign PR reviewers.

* (maint) Remove old CODEOWNERS file

This commit removes the old CODEOWNERS file. I didn't realize this existed when
adding the other one. All of Puppet's other repositories have the CODEOWNERS
file at the top level of the repo for easy findability, so that should replace
this one. This one also contained individual Puppet employees, rather than a
team, which is what we prefer.